+++ LaSiesta/MODGEN/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.1
 
 * ServerSync integration
+* Added dialog to confirm sleeping siesta
 
 ## 1.0.0